+++ laravel/app/Http/Controllers/Auth/VerifyEmailController.php
@@ -7,7 +7,7 @@ use Illuminate\Auth\Events\Verified;
 use Illuminate\Foundation\Auth\EmailVerificationRequest;
 use Illuminate\Http\RedirectResponse;
 
-class VerifyEmailController extends Controller
+class VerifyEmailController
 {
     /**
      * Mark the authenticated user's email address as verified.
