--- laravel/app/Http/Controllers/Controller.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-abstract class Controller
-{
-    //
-}
