--- laravel/resources/views/course.blade.php
@@ -1,9 +1,0 @@
-<!-- this is not used, resources/views/course/show.blade.php is used now
-<x-app-layout>
-    <div style="height: 100vh;">
-        
-        <livewire:course :courseId="$courseId" />
-        @livewireScripts
-    
-    </div>
-</x-app-layout>
