--- laravel/resources/views/discoverProfessors.blade.php
@@ -1,16 +1,0 @@
-<x-app-layout>
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Discover Professors</title>
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-        <link rel="stylesheet" href="{{ asset('css/style.css') }}">
-        @livewireStyles
-    </head>
-    
-    <body>
-        
-        <livewire:discover_professors/>
-    </body>
-</x-app-layout>
