+++ laravel/resources/views/livewire/pages/courses.blade.php
@@ -1,5 +1,4 @@
 <div>
-
     <!-- Courses Section -->
     <div class="mt-4">
         <h2 class="mb-4 text-center"> Browse Courses </h2>
