+++ laravel/resources/views/pages/courses.blade.php
@@ -1,0 +1,3 @@
+<x-app-layout>
+    @livewire('pages.courses')
+</x-app-layout>
