+++ laravel/resources/views/pages/professors.blade.php
@@ -1,0 +1,3 @@
+<x-app-layout>
+    @livewire('pages.professors')
+</x-app-layout>
