+++ laravel/resources/views/pages/search-results.blade.php
