--- laravel/resources/views/professor.blade.php
@@ -1,4 +1,0 @@
-<!-- this is not used, resources/views/professor/show.blade.php is used now
-<x-app-layout>
-    @livewire('professor')
-</x-app-layout>
